Added GRG test for area
Added a test to check the size of each polygon based on the perimeter. Using the length and height inputs  * 2 we understand the expected perimeter. The precision of the attribute does not give us an integer and because of that it can be off by one depending on rounding. Using less than or equal to one we can confirm that each grg polygon is the size we expect it to be.