Calculate range distance on output after join
The calculation to set the range value was being done on the temp table
that was to be joined to features to make the final output. For some
reason, it was not executing correctly. Changed it to do the calculation
on the final output feature class, and it is now working.