Test to see if ouput Range values match input list of values
This commit tests whether the values in the Range field of the output
matches the original list of input values for the range ring from list
test case. It tests issue 311 because the tool used to populate Range
with diameters, now it populates with radii and the lists will match.