UpdateLYRX file and supporting feature class in featuresersWebMerc.gdb
To apply symbology from a LYRX file to a new layer in Pro, the source data from that original LYRX file has to be visible.
I have copied a feature class in to the featuresetWebMerc.gdb to supply this.
The tool is now working correctly for me.
I am also seeing it add a Number field correctly when none is present or specified on the input feature class.