Update LLOS_InputObserversGDB.lyr to include feature template (#392)
* Update LLOS_InputObserversGDB.lyr to include feature template

The "LLOS_InputObserversGDB.lyr" is updated to include a feature template. This fixes an issue with the LLOS and RLOS tools in Pro 2.3 not giving a feature template when the pencil button is clicked.

* Update LLOS Input Lyr to have a feature template

Trying this layer instead.

* Update LLOS_InputObserversGDB.lyr again

Try again updating LLOS_InputObserversGDB.lyr
This time it seems to work correctly.

* Update LYR file to use relative paths

Used Save Layer File with Relative Path option to update the layer file.