Added CSV table for testing issue 386
CSV of pseudo-incidents for out AOI, with Point_X and Point_Y fields containing UTM Zone 10 N northing and easting values.
When these are used with the Table to Point tool, DD_2 (two field input) and the coordinate system paramater is set to WGS 1984 UTM Zone 10N (same as the ElevationUTM_Zone10 raster in the test data), the tool should output points in the AOI with all the attributes in the table, plus the DD_Lat and DD_Lon fields.